Aprendi como usar grupo de requisições e como fazer validações.

+++ src/main/resources/db/migration/V001__criando-tabela-cliente.sql
@@ -1,5 +1,5 @@
 create table cliente(
-	id bigint not null auto_increment,
+	id integer not null auto_increment,
     nome varchar(255) not null,
     email varchar(255) null,
     telefone varchar(255) null,
